backend/requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
groq
python-dotenv
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-core>=0.3.0
mcp[cli]>=1.0.0
httpx>=0.27.0
